+++ packages/govuk-frontend/src/govuk/helpers/colour.unit.test.js
@@ -372,3 +372,43 @@ describe('@function govuk-organisation-colour', () => {
     })
   })
 })
+
+describe('@function govuk-shade', () => {
+  it('outputs hexadecimal values', async () => {
+    const sass = `
+      @import "helpers/colour";
+
+      .foo {
+        color: govuk-shade(rgb(171, 205, 239), 17);
+      }
+    `
+
+    await expect(compileSassString(sass, sassConfig)).resolves.toMatchObject({
+      css: outdent`
+          .foo {
+            color: #8eaac6;
+          }
+        `
+    })
+  })
+})
+
+describe('@function govuk-tint', () => {
+  it('outputs hexadecimal values', async () => {
+    const sass = `
+      @import "helpers/colour";
+
+      .foo {
+        color: govuk-tint(rgb(18, 52, 86), 17);
+      }
+    `
+
+    await expect(compileSassString(sass, sassConfig)).resolves.toMatchObject({
+      css: outdent`
+          .foo {
+            color: #3a5773;
+          }
+        `
+    })
+  })
+})